% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acronym}
\alias{acronym}
\title{Acronym List}
\format{
A character vector of acronyms
}
\description{
A dataset containing common acronyms used in text processing
}
\keyword{datasets}
