% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{analyze_semantic_evolution}
\alias{analyze_semantic_evolution}
\title{Analyze Semantic Evolution}
\usage{
analyze_semantic_evolution(temporal_results, verbose = FALSE, ...)
}
\arguments{
\item{temporal_results}{Temporal analysis results}

\item{verbose}{Logical indicating whether to print progress messages}

\item{...}{Additional parameters}
}
\value{
List containing evolution analysis
}
\description{
Analyzes semantic evolution patterns in temporal results
}
\concept{topic-modeling}
