% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{calculate_assignment_consistency}
\alias{calculate_assignment_consistency}
\title{Calculate Assignment Consistency}
\usage{
calculate_assignment_consistency(assignments1, assignments2, ...)
}
\arguments{
\item{assignments1}{First set of assignments}

\item{assignments2}{Second set of assignments}

\item{...}{Additional parameters}
}
\value{
List containing consistency metrics
}
\description{
Calculates consistency between two sets of assignments
}
\concept{topic-modeling}
