% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{calculate_consistency}
\alias{calculate_consistency}
\title{Calculate Assignment Consistency}
\usage{
calculate_consistency(semantic_assignments, stm_assignments)
}
\arguments{
\item{semantic_assignments}{Semantic topic assignments.}

\item{stm_assignments}{STM topic assignments.}
}
\value{
Consistency metrics.
}
\description{
Calculates consistency between different assignment methods.
}
\keyword{internal}
