% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_docker_deployment}
\alias{check_docker_deployment}
\title{Check Docker Deployment}
\usage{
check_docker_deployment()
}
\value{
Logical TRUE if running in Docker
}
\description{
Detects whether the app is running in a Docker container.
Docker deployments have full Python/spaCy capability.
}
\keyword{internal}
