% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_web_deployment}
\alias{check_web_deployment}
\title{Check Deployment Environment}
\usage{
check_web_deployment()
}
\value{
Logical TRUE if running on web server, FALSE if local
}
\description{
Detects whether the app is running on a web server (shinyapps.io, Posit Connect)
versus locally via \code{run_app()}.
}
\keyword{internal}
