% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_similarity_matrix}
\alias{clean_similarity_matrix}
\title{Clean Similarity Matrix}
\usage{
clean_similarity_matrix(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{A numeric matrix of similarity values.}
}
\value{
Cleaned similarity matrix.
}
\description{
Cleans a similarity matrix by handling NA/Inf values,
ensuring symmetry, and setting diagonal to 1.
}
\concept{matrix-utilities}
\keyword{internal}
