% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_empty_plot_message}
\alias{create_empty_plot_message}
\title{Empty plotly placeholder}
\usage{
create_empty_plot_message(message, color = "#6B7280", font_size = 16)
}
\arguments{
\item{message}{Message to display.}

\item{color}{Text color.}

\item{font_size}{Font size in px.}
}
\value{
A \code{plotly} htmlwidget.
}
\description{
Empty plotly placeholder
}
\keyword{internal}
