% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_message_table}
\alias{create_message_table}
\title{Single-cell message DT table}
\usage{
create_message_table(message, font_size = "16px", color = "#6c757d")
}
\arguments{
\item{message}{Character message to display.}

\item{font_size}{CSS font size.}

\item{color}{CSS color.}
}
\value{
A \code{DT::datatable} htmlwidget.
}
\description{
Single-cell message DT table
}
\keyword{internal}
