% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_list_1}
\alias{dictionary_list_1}
\title{Dictionary List 1}
\format{
A character vector of dictionary terms
}
\description{
A dataset containing dictionary terms for text analysis
}
\keyword{datasets}
