% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_list_2}
\alias{dictionary_list_2}
\title{Dictionary List 2}
\format{
A character vector of dictionary terms
}
\description{
A dataset containing additional dictionary terms for text analysis
}
\keyword{datasets}
