% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{evaluate_optimal_topic_number}
\alias{evaluate_optimal_topic_number}
\title{Evaluate Optimal Number of Topics}
\usage{
evaluate_optimal_topic_number(
  dfm_object,
  K_range,
  max.em.its = 75,
  categorical_var = NULL,
  continuous_var = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{K_range}{A numeric vector specifying the range of topics (K) to search over.}

\item{max.em.its}{Maximum number of EM iterations (default: 75).}

\item{categorical_var}{An optional character string for a categorical variable in the metadata.}

\item{continuous_var}{An optional character string for a continuous variable in the metadata.}

\item{verbose}{Logical; if \code{TRUE}, prints progress information (default: TRUE).}

\item{...}{Further arguments passed to \code{stm::searchK}.}
}
\value{
A Plotly object showing the diagnostics for the number of topics (K).
}
\description{
This function performs a search for the optimal number of topics (K) using \code{stm::searchK}
and visualizes diagnostics, including held-out likelihood, residuals, semantic coherence,
and lower bound metrics.
}
\examples{
if (interactive()) {
 df <- TextAnalysisR::SpecialEduTech
 dfm_object <- df \%>\%
   TextAnalysisR::preprocess_texts(text_field = "abstract") \%>\%
   quanteda::dfm() \%>\%
   quanteda::dfm_trim(min_termfreq = 1, min_docfreq = 1) \%>\%
   .[quanteda::ntoken(.) > 0, ]
 TextAnalysisR::evaluate_optimal_topic_number(
   dfm_object = dfm_object,
   K_range = 5:30,
   max.em.its = 75,
   categorical_var = "reference_type",
   continuous_var = "year")
}

}
