% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{extract_frequent_word}
\alias{extract_frequent_word}
\title{Extract frequently observed words}
\usage{
extract_frequent_word(data, n = 20, ...)
}
\arguments{
\item{data}{A document-feature matrix (dfm) object through the quanteda package.}

\item{n}{A number of top n features (terms or words) frequently observed.}

\item{...}{Further arguments passed to \code{quanteda.textstats::textstat_frequency}.}
}
\value{
A character (feature) Vector extracted from \code{quanteda.textstats::textstat_frequency}.
The result is a character vector of top frequent words.
}
\description{
Extract frequently observed top n features (terms or words).
}
\examples{
suppressWarnings({
if(requireNamespace("quanteda")){
dfm <- SpecialEduTech \%>\%
       preprocess_texts(text_field = "abstract") \%>\%
       quanteda::dfm()
dfm \%>\% extract_frequent_word()
}
})

}
