% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dt_options}
\alias{get_dt_options}
\title{Default DT datatable options}
\usage{
get_dt_options(scroll_y = "400px", page_length = 25, show_buttons = TRUE)
}
\arguments{
\item{scroll_y}{CSS height for vertical scrolling.}

\item{page_length}{Default page length.}

\item{show_buttons}{Include copy/csv/excel/pdf/print buttons.}
}
\value{
A list for \code{DT::datatable(options = ...)}.
}
\description{
Default DT datatable options
}
\keyword{internal}
