% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gemini_embeddings}
\alias{get_gemini_embeddings}
\title{Get Gemini Embeddings (Internal)}
\usage{
get_gemini_embeddings(texts, model, api_key)
}
\description{
Get Gemini Embeddings (Internal)
}
\keyword{internal}
