% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_openai_embeddings}
\alias{get_openai_embeddings}
\title{Get OpenAI Embeddings (Internal)}
\usage{
get_openai_embeddings(texts, model, api_key)
}
\description{
Get OpenAI Embeddings (Internal)
}
\keyword{internal}
