% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_nlp.R
\name{get_spacy_nlp}
\alias{get_spacy_nlp}
\title{Get spaCy NLP Processor}
\usage{
get_spacy_nlp(model = "en_core_web_sm")
}
\arguments{
\item{model}{Model to use if not initialized}
}
\value{
The SpacyNLP Python object
}
\description{
Internal function to get the initialized spaCy processor.
}
\keyword{internal}
