% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{hierarchical_clustering}
\alias{hierarchical_clustering}
\title{Plot a Hierarchical Clustering Dendrogram}
\usage{
hierarchical_clustering(gamma_matrix, method = "ward.D2")
}
\arguments{
\item{gamma_matrix}{A document-topic probability matrix (output of \code{stm_model$gamma}).}

\item{method}{Clustering method (default is "ward.D2").}
}
\value{
A ggplot2 dendrogram plot.
}
\description{
Perform hierarchical clustering on STM model's document-topic probabilities
and plot a dendrogram.
}
\examples{
if (interactive()) {
  # Assume stm_model is a fitted STM model.
  gamma_matrix <- stm_model$gamma
  hierarchical_clustering(gamma_matrix, method = "ward.D2")
}

}
