% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{identify_topic_trends}
\alias{identify_topic_trends}
\title{Identify Topic Trends}
\usage{
identify_topic_trends(temporal_results, ...)
}
\arguments{
\item{temporal_results}{Temporal analysis results}

\item{...}{Additional parameters}
}
\value{
List containing identified trends
}
\description{
Identifies trending topics in temporal results
}
\concept{topic-modeling}
