% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{lexical_analysis}
\alias{lexical_analysis}
\title{Lexical Analysis Functions}
\description{
Functions for lexical analysis including:
\itemize{
\item Linguistic Annotation (POS tagging, NER)
\item Frequency Analysis (word frequency, n-grams, MWEs)
\item Keywords (TF-IDF, keyness)
\item Lexical Diversity (TTR, MTLD, MATTR)
\item Readability (Flesch, Gunning Fog, etc.)
}
}
\concept{lexical}
