% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_mean_topic_prevalence}
\alias{plot_mean_topic_prevalence}
\title{Plot Mean Topic Prevalence Across Documents}
\usage{
plot_mean_topic_prevalence(
  dfm_object,
  K_number,
  max.em.its = 75,
  categorical_var = NULL,
  continuous_var = NULL,
  top_n = 15,
  height = 500,
  width = 1000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{K_number}{The number of topics to display.}

\item{max.em.its}{Maximum number of EM iterations (default: 75).}

\item{categorical_var}{An optional character string for a categorical variable in the metadata.}

\item{continuous_var}{An optional character string for a continuous variable in the metadata.}

\item{top_n}{The number of topics to display, ordered by their mean prevalence.}

\item{height}{The height of the resulting Plotly plot, in pixels. Defaults to \code{500}.}

\item{width}{The width of the resulting Plotly plot, in pixels. Defaults to \code{1000}.}

\item{verbose}{Logical; if \code{TRUE}, prints progress information (default: TRUE).}

\item{...}{Further arguments passed to \code{stm::searchK}.}
}
\value{
A \code{ggplot} object showing a bar plot of topic prevalence. Topics are ordered by their
mean gamma value (average prevalence across documents).
}
\description{
This function calculates the mean topic prevalence across documents and plots the top topics.
}
\examples{
if (interactive()) {
 df <- TextAnalysisR::SpecialEduTech
 dfm_object <- df \%>\%
   TextAnalysisR::preprocess_texts(text_field = "abstract") \%>\%
   quanteda::dfm() \%>\%
   quanteda::dfm_trim(min_termfreq = 1, min_docfreq = 1) \%>\%
   .[quanteda::ntoken(.) > 0, ]
TextAnalysisR::plot_mean_topic_prevalence(
  dfm_object = dfm_object,
  K_number = 15,
  max.em.its = 75,
  categorical_var = "reference_type",
  continuous_var = "year",
  top_n = 15,
  height = 500,
  width = 1000,
  verbose = TRUE)
}

}
