% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_topic_probability_table}
\alias{plot_topic_probability_table}
\title{Visualize a table for document-topic probabilities}
\usage{
plot_topic_probability_table(data, top_n, topic_names = NULL, ...)
}
\arguments{
\item{data}{A tidy data frame that includes document-topic probabilities.
(probabilities of each topic per each document).}

\item{top_n}{A number of top n terms with highest document-topic probabilities.}

\item{topic_names}{Topic names}

\item{...}{Further arguments passed.}
}
\value{
A tibble (data frame) object with a list of topic probabilities from \code{tidytext::tidy}.
The result is a ggplot object representing the topic-term plot.
}
\description{
Visualize a table for document-topic probabilities.
}
\examples{
if(requireNamespace("quanteda", "tidytext")){
dfm <- SpecialEduTech \%>\% preprocess_texts(text_field = "abstract") \%>\% quanteda::dfm()
data <- tidytext::tidy(stm_15, matrix = "gamma", document_names = rownames(dfm), log = FALSE)
data \%>\% plot_topic_probability_table(top_n = 15)
}

}
