% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_word_correlation_network}
\alias{plot_word_correlation_network}
\title{Plot a Word Correlation Network}
\usage{
plot_word_correlation_network(
  dfm_object,
  co_occur_n = 30,
  corr_n = 0.4,
  height = 900,
  width = 800
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{co_occur_n}{Minimum number of co-occurrences for filtering terms (default is 30).}

\item{corr_n}{Minimum correlation value for filtering terms (default is 0.4).}

\item{height}{The height of the resulting Plotly plot, in pixels. Defaults to \code{900}.}

\item{width}{The width of the resulting Plotly plot, in pixels. Defaults to \code{800}.}
}
\value{
A Plotly object visualizing the interactive word correlation network.
}
\description{
Visualize the correlation relationships between terms in the corpus based on pairwise correlations.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
  TextAnalysisR::plot_word_correlation_network(
    dfm_object,
    co_occur_n = 30,
    corr_n = 0.4,
    height = 900,
    width = 800)
}
}
