% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{preprocess_texts}
\alias{preprocess_texts}
\title{Preprocess Text Data}
\usage{
preprocess_texts(
  united_tbl,
  text_field = "united_texts",
  min_char = 2,
  remove_punct = TRUE,
  remove_symbols = TRUE,
  remove_numbers = TRUE,
  remove_url = TRUE,
  remove_separators = TRUE,
  split_hyphens = TRUE,
  split_tags = TRUE,
  include_docvars = TRUE,
  keep_acronyms = FALSE,
  padding = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{united_tbl}{A data frame that contains text data.}

\item{text_field}{The name of the column that contains the text data.}

\item{min_char}{The minimum number of characters for a token to be included (default: 2).}

\item{remove_punct}{Logical; remove punctuation from the text (default: TRUE).}

\item{remove_symbols}{Logical; remove symbols from the text (default: TRUE).}

\item{remove_numbers}{Logical; remove numbers from the text (default: TRUE).}

\item{remove_url}{Logical; remove URLs from the text (default: TRUE).}

\item{remove_separators}{Logical; remove separators from the text (default: TRUE).}

\item{split_hyphens}{Logical; split hyphenated words into separate tokens (default: TRUE).}

\item{split_tags}{Logical; split tags into separate tokens (default: TRUE).}

\item{include_docvars}{Logical; include document variables in the tokens object (default: TRUE).}

\item{keep_acronyms}{Logical; keep acronyms in the text (default: FALSE).}

\item{padding}{Logical; add padding to the tokens object (default: FALSE).}

\item{verbose}{Logical; print verbose output (default: FALSE).}

\item{...}{Additional arguments passed to \code{quanteda::tokens}.}
}
\value{
A \code{tokens} object that contains the preprocessed text data.
}
\description{
Preprocesses text data by:
\itemize{
\item Constructing a corpus
\item Tokenizing text into words
\item Converting to lowercase
\item Specifying a minimum token length.
}

Typically used before constructing a dfm and fitting an STM model.
}
\examples{
if (interactive()) {
df <- TextAnalysisR::SpecialEduTech

united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))

tokens <- TextAnalysisR::preprocess_texts(united_tbl,
                                         text_field = "united_texts",
                                         min_char = 2,
                                         remove_punct = TRUE,
                                         remove_symbols = TRUE,
                                         remove_numbers = TRUE,
                                         remove_url = TRUE,
                                         remove_separators = TRUE,
                                         split_hyphens = TRUE,
                                         split_tags = TRUE,
                                         include_docvars = TRUE,
                                         keep_acronyms = FALSE,
                                         padding = FALSE,
                                         verbose = FALSE)
print(tokens)
}
}
