% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{process_files}
\alias{process_files}
\title{Process Files}
\usage{
process_files(dataset_choice, file_info = NULL, text_input = NULL)
}
\arguments{
\item{dataset_choice}{A character string indicating the dataset choice.}

\item{file_info}{A data frame containing file information with a column named 'filepath' (default: NULL).}

\item{text_input}{A character string containing text input (default: NULL).}
}
\value{
A data frame containing the processed data.
}
\description{
This function processes different types of files and text input based on the dataset choice.
}
\examples{
if (interactive()) {
  mydata <- TextAnalysisR::process_files(dataset_choice = "Upload an Example Dataset")
  head(mydata)

  file_info <- data.frame(filepath = "inst/extdata/SpecialEduTech.xlsx")
  mydata <- TextAnalysisR::process_files(dataset_choice = "Upload Your File",
                                          file_info = file_info)
  head(mydata)

  text_input <- paste0("The purpose of this study was to conduct a content analysis of ",
                       "research on technology use.")
  mydata <- TextAnalysisR::process_files(dataset_choice = "Copy and Paste Text",
                                          text_input = text_input)
  head(mydata)
}
}
