% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{remove_common_words}
\alias{remove_common_words}
\title{Remove Common Words Across Documents}
\usage{
remove_common_words(tokens, remove_vars, dfm_object)
}
\arguments{
\item{tokens}{A \code{tokens} object from the \code{quanteda} package, typically processed
using functions like \code{tokens_select} or \code{tokens_remove}.}

\item{remove_vars}{A character vector of words to remove from the tokens. If \code{NULL},
the function returns the result of \code{dfm_init_func()}.}

\item{dfm_object}{A \code{dfm} object to process after removing the specified words.}
}
\value{
A \code{dfm} object with the specified words removed and the remaining tokens categorized
}
\description{
This function removes specified common words from a tokens object and applies two dictionaries
to categorize the remaining tokens. It returns a document-feature matrix (dfm) based on the
processed tokens. If no words are specified for removal, it returns an initial dfm using the
provided initialization function.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
  TextAnalysisR::remove_common_words(tokens = tokens,
                                     remove_vars = c("level", "testing"),
                                     dfm_object = dfm_object)
}
}
