% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_notification_by_id}
\alias{remove_notification_by_id}
\title{Remove Shiny notification by id}
\usage{
remove_notification_by_id(id, session = NULL)
}
\arguments{
\item{id}{Notification id.}

\item{session}{Shiny session. Defaults to the current reactive domain.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Remove Shiny notification by id
}
\keyword{internal}
