% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renumber_clusters_sequentially}
\alias{renumber_clusters_sequentially}
\title{Renumber Clusters Sequentially}
\usage{
renumber_clusters_sequentially(clusters)
}
\arguments{
\item{clusters}{A vector of cluster assignments.}
}
\value{
Vector with clusters renumbered sequentially (1, 2, 3, ...).
}
\description{
Renumbers cluster assignments to sequential integers
starting from 1.
}
\concept{matrix-utilities}
\keyword{internal}
