% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_analysis.R
\name{semantic_document_clustering}
\alias{semantic_document_clustering}
\title{Semantic Document Clustering}
\usage{
semantic_document_clustering(embeddings, method = "UMAP", clusters = NULL, ...)
}
\arguments{
\item{embeddings}{Document embeddings matrix}

\item{method}{Dimensionality reduction method ("PCA", "t-SNE", "UMAP")}

\item{clusters}{Optional cluster assignments}

\item{...}{Additional arguments}
}
\value{
A ggplot2 visualization of document clusters
}
\description{
Creates a unified visualization of document clustering with optional clustering
}
\concept{semantic}
