% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_dfm_instructions_modal}
\alias{show_dfm_instructions_modal}
\title{Show DFM instructions modal}
\usage{
show_dfm_instructions_modal(
  output_id,
  feature_name = "this feature",
  session = NULL
)
}
\arguments{
\item{output_id}{Output id to reset after modal closes.}

\item{feature_name}{Feature description in the instructions.}

\item{session}{Shiny session.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Show DFM instructions modal
}
\keyword{internal}
