% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_guide_modal}
\alias{show_guide_modal}
\title{Show guide modal}
\usage{
show_guide_modal(guide_name, title, size = "l")
}
\arguments{
\item{guide_name}{Guide identifier (filename stem).}

\item{title}{Modal title.}

\item{size}{Modal size ("s", "m", "l", "xl").}
}
\value{
Invisibly \code{NULL}.
}
\description{
Show guide modal
}
\keyword{internal}
