% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_no_dfm_notification}
\alias{show_no_dfm_notification}
\title{Show no-DFM notification}
\usage{
show_no_dfm_notification(feature_name = "this feature", duration = 7)
}
\arguments{
\item{feature_name}{Feature description in the message.}

\item{duration}{Seconds until auto-dismiss.}
}
\value{
Invisibly the notification id.
}
\description{
Show no-DFM notification
}
\keyword{internal}
