% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_no_feature_matrix_notification}
\alias{show_no_feature_matrix_notification}
\title{Show no-feature-matrix notification}
\usage{
show_no_feature_matrix_notification(duration = 7)
}
\arguments{
\item{duration}{Seconds until auto-dismiss.}
}
\value{
Invisibly the notification id.
}
\description{
Show no-feature-matrix notification
}
\keyword{internal}
