% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_unite_texts_required_notification}
\alias{show_unite_texts_required_notification}
\title{Show united-texts-required notification}
\usage{
show_unite_texts_required_notification(duration = 5)
}
\arguments{
\item{duration}{Seconds until auto-dismiss.}
}
\value{
Invisibly the notification id.
}
\description{
Show united-texts-required notification
}
\keyword{internal}
