% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stm_15}
\alias{stm_15}
\title{An example structure of a structural topic model}
\format{
An object of class \code{STM} of length 11.
}
\usage{
stm_15
}
\description{
Contains 15 topics, topic prevalences, and etc. from \code{stm::stm}.
}
\examples{
stm_15

}
\keyword{datasets}
