% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_list}
\alias{stopwords_list}
\title{Stopwords List}
\format{
A character vector of stopwords
}
\description{
A dataset containing stopwords for text preprocessing
}
\keyword{datasets}
