% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysisR.app.R
\name{TextAnalysisR.app}
\alias{TextAnalysisR.app}
\title{Launch and browse the TextAnalysisR app}
\usage{
TextAnalysisR.app()
}
\description{
Launch and browse the TextAnalysisR app.
}
\examples{
if (interactive()) {
  library(TextAnalysisR)
  TextAnalysisR.app()
}

}
