% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{unite_text_cols}
\alias{unite_text_cols}
\title{Unite Text Columns}
\usage{
unite_text_cols(df, listed_vars)
}
\arguments{
\item{df}{A data frame that contains text data.}

\item{listed_vars}{A character vector of column names to be united into "united_texts".}
}
\value{
A data frame with a new column "united_texts" created by uniting the specified variables.
}
\description{
This function unites specified text columns in a data frame into a single column named "united_texts" while retaining the original columns.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech

  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  print(united_tbl)
}
}
