% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{validate_semantic_coherence}
\alias{validate_semantic_coherence}
\title{Validate Semantic Coherence}
\usage{
validate_semantic_coherence(embeddings, topic_assignments, ...)
}
\arguments{
\item{embeddings}{Document embeddings matrix.}

\item{topic_assignments}{Vector of topic assignments for documents.}

\item{...}{Additional parameters (currently unused).}
}
\value{
List containing coherence score and metrics.
}
\description{
Validates the semantic coherence of topic assignments using
intra-cluster distance in embedding space.
}
\concept{topic-modeling}
