% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{word_frequency_trends}
\alias{word_frequency_trends}
\title{Visualize Word Frequency Trends Over Time}
\usage{
word_frequency_trends(dfm_object, gamma_td, time_variable, selected_terms)
}
\arguments{
\item{dfm_object}{A \code{quanteda} dfm object representing document-feature matrix.}

\item{gamma_td}{A tidy data frame of document-topic probabilities (gamma).}

\item{time_variable}{The column name for the time variable (e.g., "year").}

\item{selected_terms}{A vector of terms to analyze trends for.}
}
\value{
A ggplot2 line plot showing word frequency trends.

A ggplot object showing word frequency trends over time.
}
\description{
Analyze and visualize the change in word frequencies over time or any continuous variable.

Analyze and visualize word frequency trends over time for a fixed term column.
}
\examples{
if (interactive()) {
df <- TextAnalysisR::SpecialEduTech
dfm_object <- preprocess_texts(df, text_field = "abstract") \%>\%
 quanteda::dfm()
word_frequency_trends(dfm_object, gamma_td,
                     time_variable = "year",
                     selected_terms = c("computer", "disability"))
}

if (interactive()) {
# Assume stm_model is a fitted STM model.
df <- TextAnalysisR::SpecialEduTech
dfm_object <- preprocess_texts(df, text_field = "abstract") \%>\%
 quanteda::dfm()
gamma_td <- tidytext::tidy(stm_model, matrix = "gamma")
word_frequency_trends(dfm_object,
                     gamma_td, time_variable = "year",
                     selected_terms = c("computer", "disability"))
}

}
