% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_text_for_tooltip}
\alias{wrap_text_for_tooltip}
\title{Wrap text for tooltip display}
\usage{
wrap_text_for_tooltip(
  text,
  max_words = 150,
  chars_per_line = 50,
  max_lines = 3
)
}
\arguments{
\item{text}{Input string.}

\item{max_words}{Maximum words to keep.}

\item{chars_per_line}{Approximate line width.}

\item{max_lines}{Maximum lines.}
}
\value{
A string with HTML \verb{<br>} breaks.
}
\description{
Wrap text for tooltip display
}
\keyword{internal}
