% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acronym}
\alias{acronym}
\title{Acronym list}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 145 rows and 2 columns.
}
\usage{
acronym
}
\description{
Contains acronyms commonly observed in the bibliographic data of research on the use of technology in teaching mathematics to students with disabilities.
}
\examples{
acronym

}
\keyword{datasets}
