% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{analyze_semantic_evolution}
\alias{analyze_semantic_evolution}
\title{Analyze Semantic Evolution}
\usage{
analyze_semantic_evolution(temporal_results, verbose = FALSE, ...)
}
\arguments{
\item{temporal_results}{Temporal analysis results}

\item{verbose}{Logical indicating whether to print progress messages}

\item{...}{Additional parameters}
}
\value{
List containing evolution analysis
}
\description{
Analyzes semantic evolution patterns in temporal results
}
\seealso{
Other topic-modeling: 
\code{\link{assess_embedding_stability}()},
\code{\link{assess_hybrid_stability}()},
\code{\link{auto_tune_embedding_topics}()},
\code{\link{calculate_assignment_consistency}()},
\code{\link{calculate_keyword_stability}()},
\code{\link{calculate_npmi}()},
\code{\link{calculate_semantic_drift}()},
\code{\link{calculate_topic_diversity}()},
\code{\link{calculate_topic_probability}()},
\code{\link{calculate_topic_stability}()},
\code{\link{extract_topic_terms_df}()},
\code{\link{find_optimal_k}()},
\code{\link{find_topic_matches}()},
\code{\link{fit_embedding_model}()},
\code{\link{fit_hybrid_model}()},
\code{\link{fit_temporal_model}()},
\code{\link{fit_topic_prevalence_model}()},
\code{\link{generate_topic_labels}()},
\code{\link{get_topic_prevalence}()},
\code{\link{get_topic_terms}()},
\code{\link{get_topic_texts}()},
\code{\link{identify_topic_trends}()},
\code{\link{plot_model_comparison}()},
\code{\link{plot_quality_metrics}()},
\code{\link{plot_topic_effects_categorical}()},
\code{\link{plot_topic_effects_continuous}()},
\code{\link{plot_topic_probability}()},
\code{\link{plot_word_probability}()},
\code{\link{run_neural_topics_internal}()},
\code{\link{validate_semantic_coherence}()}
}
\concept{topic-modeling}
