% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{calculate_assignment_consistency}
\alias{calculate_assignment_consistency}
\title{Calculate Assignment Consistency}
\usage{
calculate_assignment_consistency(assignments1, assignments2, ...)
}
\arguments{
\item{assignments1}{First set of assignments}

\item{assignments2}{Second set of assignments}

\item{...}{Additional parameters}
}
\value{
List containing consistency metrics
}
\description{
Calculates consistency between two sets of assignments
}
\seealso{
Other topic-modeling: 
\code{\link{analyze_semantic_evolution}()},
\code{\link{assess_hybrid_stability}()},
\code{\link{calculate_eval_metrics_internal}()},
\code{\link{calculate_keyword_stability}()},
\code{\link{calculate_semantic_drift}()},
\code{\link{calculate_topic_probability}()},
\code{\link{calculate_topic_stability}()},
\code{\link{find_optimal_k}()},
\code{\link{find_topic_matches}()},
\code{\link{fit_embedding_topics}()},
\code{\link{fit_hybrid_model}()},
\code{\link{fit_temporal_model}()},
\code{\link{generate_topic_labels}()},
\code{\link{get_topic_prevalence}()},
\code{\link{get_topic_terms}()},
\code{\link{get_topic_texts}()},
\code{\link{identify_topic_trends}()},
\code{\link{plot_model_comparison}()},
\code{\link{plot_quality_metrics}()},
\code{\link{run_contrastive_topics_internal}()},
\code{\link{run_neural_topics_internal}()},
\code{\link{run_temporal_topics_internal}()},
\code{\link{validate_semantic_coherence}()}
}
\concept{topic-modeling}
