% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_cosine_similarity}
\alias{calculate_cosine_similarity}
\title{Calculate Cosine Similarity Matrix}
\usage{
calculate_cosine_similarity(matrix_data)
}
\arguments{
\item{matrix_data}{A numeric matrix where rows represent documents/observations}
}
\value{
A square similarity matrix with values between -1 and 1
}
\description{
Calculates the cosine similarity between all pairs of rows in a matrix.
}
