% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{calculate_keyword_stability}
\alias{calculate_keyword_stability}
\title{Calculate Keyword Stability}
\usage{
calculate_keyword_stability(keywords1, keywords2)
}
\arguments{
\item{keywords1}{First set of keywords.}

\item{keywords2}{Second set of keywords.}
}
\value{
Stability score.
}
\description{
Calculates stability between two sets of topic keywords.
}
\seealso{
Other topic-modeling:
\code{\link[=analyze_semantic_evolution]{analyze_semantic_evolution()}},
\code{\link[=assess_embedding_stability]{assess_embedding_stability()}},
\code{\link[=auto_tune_embedding_topics]{auto_tune_embedding_topics()}},
\code{\link[=calculate_assignment_consistency]{calculate_assignment_consistency()}},
\code{\link[=calculate_npmi]{calculate_npmi()}},
\code{\link[=calculate_semantic_drift]{calculate_semantic_drift()}},
\code{\link[=calculate_topic_diversity]{calculate_topic_diversity()}},
\code{\link[=calculate_topic_probability]{calculate_topic_probability()}},
\code{\link[=calculate_topic_stability]{calculate_topic_stability()}},
\code{\link[=extract_topic_terms_df]{extract_topic_terms_df()}},
\code{\link[=find_optimal_k]{find_optimal_k()}},
\code{\link[=find_topic_matches]{find_topic_matches()}},
\code{\link[=fit_embedding_model]{fit_embedding_model()}},
\code{\link[=fit_temporal_model]{fit_temporal_model()}},
\code{\link[=fit_topic_prevalence_model]{fit_topic_prevalence_model()}},
\code{\link[=generate_topic_labels]{generate_topic_labels()}},
\code{\link[=get_topic_prevalence]{get_topic_prevalence()}},
\code{\link[=get_topic_terms]{get_topic_terms()}},
\code{\link[=get_topic_texts]{get_topic_texts()}},
\code{\link[=identify_topic_trends]{identify_topic_trends()}},
\code{\link[=plot_model_comparison]{plot_model_comparison()}},
\code{\link[=plot_quality_metrics]{plot_quality_metrics()}},
\code{\link[=plot_topic_effects_categorical]{plot_topic_effects_categorical()}},
\code{\link[=plot_topic_effects_continuous]{plot_topic_effects_continuous()}},
\code{\link[=plot_topic_probability]{plot_topic_probability()}},
\code{\link[=plot_word_probability]{plot_word_probability()}},
\code{\link[=run_neural_topics_internal]{run_neural_topics_internal()}},
\code{\link[=validate_semantic_coherence]{validate_semantic_coherence()}}
}
\concept{topic-modeling}
