% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{calculate_keyword_stability}
\alias{calculate_keyword_stability}
\title{Calculate Keyword Stability}
\usage{
calculate_keyword_stability(keywords1, keywords2)
}
\arguments{
\item{keywords1}{First set of keywords.}

\item{keywords2}{Second set of keywords.}
}
\value{
Stability score.
}
\description{
Calculates stability between two sets of topic keywords.
}
