% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{clear_lexdiv_cache}
\alias{clear_lexdiv_cache}
\title{Clear Lexical Diversity Cache}
\usage{
clear_lexdiv_cache()
}
\value{
Invisible NULL
}
\description{
Clears the internal cache used for lexical diversity calculations.
Call this function if you need to free memory or ensure fresh calculations.
}
\seealso{
Other lexical: 
\code{\link{calculate_text_readability}()},
\code{\link{detect_multi_words}()},
\code{\link{extract_keywords_keyness}()},
\code{\link{extract_keywords_tfidf}()},
\code{\link{extract_morphology}()},
\code{\link{extract_named_entities}()},
\code{\link{extract_pos_tags}()},
\code{\link{lexical_analysis}},
\code{\link{lexical_diversity_analysis}()},
\code{\link{lexical_frequency_analysis}()},
\code{\link{plot_keyness_keywords}()},
\code{\link{plot_keyword_comparison}()},
\code{\link{plot_lexical_diversity_distribution}()},
\code{\link{plot_morphology_feature}()},
\code{\link{plot_readability_by_group}()},
\code{\link{plot_readability_distribution}()},
\code{\link{plot_tfidf_keywords}()},
\code{\link{plot_top_readability_documents}()},
\code{\link{render_displacy_dep}()},
\code{\link{render_displacy_ent}()},
\code{\link{summarize_morphology}()}
}
\concept{lexical}
