% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{evaluate_optimal_topic_number}
\alias{evaluate_optimal_topic_number}
\title{Evaluate Optimal Number of Topics}
\usage{
evaluate_optimal_topic_number(
  dfm_object,
  topic_range,
  max.em.its = 75,
  categorical_var = NULL,
  continuous_var = NULL,
  height = 600,
  width = 800,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dfm_object}{A \code{quanteda} document-feature matrix (dfm).}

\item{topic_range}{A numeric vector specifying the range of topics (K) to search over.}

\item{max.em.its}{Maximum number of EM iterations (default: 75).}

\item{categorical_var}{An optional character string for a categorical variable in the metadata.}

\item{continuous_var}{An optional character string for a continuous variable in the metadata.}

\item{height}{The height of the resulting Plotly plot in pixels (default: 600).}

\item{width}{The width of the resulting Plotly plot in pixels (default: 800).}

\item{verbose}{Logical; if \code{TRUE}, prints progress information.}

\item{...}{Further arguments passed to \code{stm::searchK}.}
}
\value{
A \code{plotly} object showing the diagnostics for the number of topics (K).
}
\description{
This function performs a search for the optimal number of topics (K) using \code{stm::searchK}
and visualizes diagnostics, including held-out likelihood, residuals, semantic coherence,
and lower bound metrics.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
  TextAnalysisR::evaluate_optimal_topic_number(
    dfm_object = dfm_object,
    topic_range = 5:30,
    max.em.its = 75,
    categorical_var = "reference_type",
    continuous_var = "year",
    height = 600,
    width = 800,
    verbose = TRUE
  )
}
}
