% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{examine_top_terms}
\alias{examine_top_terms}
\title{Examine highest per-term per-topic probabilities}
\usage{
examine_top_terms(data, top_n, ...)
}
\arguments{
\item{data}{A tidy data frame that includes per-term per-topic probabilities (beta).}

\item{top_n}{A number of highest per-term per-topic probabilities in each document (number of top_n can be changed).}

\item{...}{Further arguments passed to \code{dplyr::group_by}.}
}
\value{
A tibble (data frame) object with a list of word probabilities from \code{tidytext::tidy}.
The result is a data frame containing word probabilities for each topic.
}
\description{
Examine highest per-term per-topic probabilities.
}
\examples{
suppressWarnings({
if(requireNamespace("quanteda", "tidytext")){
dfm <- SpecialEduTech \%>\%
       preprocess_texts(text_field = "abstract") \%>\%
       quanteda::dfm()
data <- tidytext::tidy(stm_15, document_names = rownames(dfm), log = FALSE)
data \%>\% examine_top_terms(top_n = 5) \%>\%
dplyr::mutate_if(is.numeric, ~ round(., 3)) \%>\%
DT::datatable(rownames = FALSE)
}
})

}
