% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{extract_keywords_tfidf}
\alias{extract_keywords_tfidf}
\title{Extract Keywords Using TF-IDF}
\usage{
extract_keywords_tfidf(dfm, top_n = 20, normalize = FALSE)
}
\arguments{
\item{dfm}{A quanteda dfm object}

\item{top_n}{Number of top keywords to extract (default: 20)}

\item{normalize}{Logical, whether to normalize TF-IDF scores to 0-1 range (default: FALSE)}
}
\value{
Data frame with columns: Keyword, TF_IDF_Score, Frequency
}
\description{
Extracts top keywords from a document-feature matrix using TF-IDF weighting.
}
\examples{
\dontrun{
library(quanteda)
corp <- corpus(c("text analysis", "data mining", "text mining"))
dfm_obj <- dfm(tokens(corp))
keywords <- extract_keywords_tfidf(dfm_obj, top_n = 5)
print(keywords)
}
}
\seealso{
Other lexical: 
\code{\link{calculate_text_readability}()},
\code{\link{clear_lexdiv_cache}()},
\code{\link{detect_multi_words}()},
\code{\link{extract_keywords_keyness}()},
\code{\link{extract_named_entities}()},
\code{\link{extract_pos_tags}()},
\code{\link{lexical_analysis}},
\code{\link{lexical_diversity_analysis}()},
\code{\link{lexical_frequency_analysis}()},
\code{\link{plot_keyness_keywords}()},
\code{\link{plot_keyword_comparison}()},
\code{\link{plot_lexical_diversity_distribution}()},
\code{\link{plot_readability_by_group}()},
\code{\link{plot_readability_distribution}()},
\code{\link{plot_tfidf_keywords}()},
\code{\link{plot_top_readability_documents}()}
}
\concept{lexical}
