% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{find_optimal_k}
\alias{find_optimal_k}
\title{Find Optimal Number of Topics}
\usage{
find_optimal_k(
  dfm_object,
  topic_range,
  max.em.its = 75,
  categorical_var = NULL,
  continuous_var = NULL,
  height = 600,
  width = 800,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dfm_object}{A quanteda dfm object to be used for topic modeling.}

\item{topic_range}{A vector of K values to test (e.g., 2:10).}

\item{max.em.its}{Maximum number of EM iterations (default: 75).}

\item{categorical_var}{Optional categorical variable(s) for prevalence.}

\item{continuous_var}{Optional continuous variable(s) for prevalence.}

\item{height}{Plot height in pixels (default: 600).}

\item{width}{Plot width in pixels (default: 800).}

\item{verbose}{Logical indicating whether to print progress (default: TRUE).}

\item{...}{Additional arguments passed to stm::searchK.}
}
\value{
A list containing search results and diagnostic plots.
}
\description{
Searches for the optimal number of topics (K) using stm::searchK.
Produces diagnostic plots to help select the best K value.
}
