% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{find_similar_words}
\alias{find_similar_words}
\title{Find Similar Words}
\usage{
find_similar_words(word, top_n = 10L, model = "en_core_web_md")
}
\arguments{
\item{word}{Character; target word.}

\item{top_n}{Integer; number of similar words to return (default: 10).}

\item{model}{Character; spaCy model to use (default: "en_core_web_md").}
}
\value{
A data frame with similar words and similarity scores.
}
\description{
Find words most similar to a given word using word vectors.
Requires a spaCy model with word vectors (en_core_web_md or en_core_web_lg).
}
\seealso{
Other lexical: 
\code{\link{calculate_dispersion_metrics}()},
\code{\link{calculate_lexical_dispersion}()},
\code{\link{calculate_log_odds_ratio}()},
\code{\link{calculate_text_readability}()},
\code{\link{clear_lexdiv_cache}()},
\code{\link{detect_multi_words}()},
\code{\link{extract_keywords_keyness}()},
\code{\link{extract_keywords_tfidf}()},
\code{\link{extract_morphology}()},
\code{\link{extract_named_entities}()},
\code{\link{extract_noun_chunks}()},
\code{\link{extract_pos_tags}()},
\code{\link{extract_subjects_objects}()},
\code{\link{get_sentences}()},
\code{\link{get_spacy_embeddings}()},
\code{\link{get_spacy_model_info}()},
\code{\link{get_word_similarity}()},
\code{\link{init_spacy_nlp}()},
\code{\link{lexical_analysis}},
\code{\link{lexical_diversity_analysis}()},
\code{\link{lexical_frequency_analysis}()},
\code{\link{parse_morphology_string}()},
\code{\link{plot_keyness_keywords}()},
\code{\link{plot_keyword_comparison}()},
\code{\link{plot_lexical_diversity_distribution}()},
\code{\link{plot_morphology_feature}()},
\code{\link{plot_readability_by_group}()},
\code{\link{plot_readability_distribution}()},
\code{\link{plot_tfidf_keywords}()},
\code{\link{plot_top_readability_documents}()},
\code{\link{render_displacy_dep}()},
\code{\link{render_displacy_ent}()},
\code{\link{spacy_extract_entities}()},
\code{\link{spacy_has_vectors}()},
\code{\link{spacy_initialized}()},
\code{\link{spacy_lemmatize}()},
\code{\link{spacy_parse_full}()},
\code{\link{summarize_morphology}()}
}
\concept{lexical}
