% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_analysis.R
\name{generate_cluster_labels_auto}
\alias{generate_cluster_labels_auto}
\title{Generate Cluster Labels}
\usage{
generate_cluster_labels_auto(
  feature_matrix,
  clusters,
  method = "tfidf",
  n_terms = 3
)
}
\arguments{
\item{feature_matrix}{Feature matrix used for clustering}

\item{clusters}{Cluster assignments}

\item{method}{Label generation method ("tfidf", "representative", "frequent")}

\item{n_terms}{Number of terms per label}
}
\value{
Named list of cluster labels
}
\description{
Generate descriptive labels for document clusters
}
