% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_analysis.R
\name{generate_cluster_labels_auto}
\alias{generate_cluster_labels_auto}
\title{Generate Cluster Labels}
\usage{
generate_cluster_labels_auto(
  feature_matrix,
  clusters,
  method = "tfidf",
  n_terms = 3
)
}
\arguments{
\item{feature_matrix}{Feature matrix used for clustering}

\item{clusters}{Cluster assignments}

\item{method}{Label generation method ("tfidf", "representative", "frequent")}

\item{n_terms}{Number of terms per label}
}
\value{
Named list of cluster labels
}
\description{
Generate descriptive labels for document clusters
}
\seealso{
Other semantic: 
\code{\link{analyze_document_clustering}()},
\code{\link{analyze_similarity_gaps}()},
\code{\link{calculate_clustering_metrics}()},
\code{\link{calculate_cross_similarity}()},
\code{\link{calculate_document_similarity}()},
\code{\link{calculate_similarity_robust}()},
\code{\link{cluster_embeddings}()},
\code{\link{cross_analysis_validation}()},
\code{\link{export_document_clustering}()},
\code{\link{extract_cross_category_similarities}()},
\code{\link{fit_semantic_model}()},
\code{\link{generate_cluster_labels}()},
\code{\link{generate_embeddings}()},
\code{\link{reduce_dimensions}()},
\code{\link{semantic_document_clustering}()},
\code{\link{semantic_similarity_analysis}()},
\code{\link{temporal_semantic_analysis}()},
\code{\link{validate_cross_models}()},
\code{\link{word_co_occurrence_network}()},
\code{\link{word_correlation_network}()}
}
\concept{semantic}
