% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{generate_topic_keywords}
\alias{generate_topic_keywords}
\title{Generate Topic Keywords}
\usage{
generate_topic_keywords(texts, topic_assignments, n_keywords = 10)
}
\arguments{
\item{texts}{A character vector of texts.}

\item{topic_assignments}{A vector of topic assignments.}

\item{n_keywords}{The number of keywords to extract per topic.}
}
\value{
A list of keywords for each topic.
}
\description{
Internal function to generate keywords for topics using TF-IDF analysis.
}
\keyword{internal}
