% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dt_options}
\alias{get_dt_options}
\title{Get Standard DataTable Options}
\usage{
get_dt_options(scroll_y = "400px", page_length = 25, show_buttons = TRUE)
}
\arguments{
\item{scroll_y}{Vertical scroll height (default: "400px")}

\item{page_length}{Number of rows per page (default: 25)}

\item{show_buttons}{Whether to show export buttons (default: TRUE}
}
\value{
A list of DT options
}
\description{
Returns standardized DT::datatable options for consistent table formatting
across the TextAnalysisR application.
}
\examples{
\dontrun{
DT::datatable(my_data, options = get_dt_options())
DT::datatable(my_data, options = get_dt_options(scroll_y = "300px"))
}
}
\seealso{
Other visualization: 
\code{\link{apply_standard_plotly_layout}()},
\code{\link{create_empty_plot_message}()},
\code{\link{create_message_table}()},
\code{\link{create_standard_ggplot_theme}()},
\code{\link{get_plotly_hover_config}()},
\code{\link{get_sentiment_color}()},
\code{\link{get_sentiment_colors}()},
\code{\link{plot_cluster_terms}()},
\code{\link{plot_cross_category_heatmap}()},
\code{\link{plot_entity_frequencies}()},
\code{\link{plot_mwe_frequency}()},
\code{\link{plot_ngram_frequency}()},
\code{\link{plot_pos_frequencies}()},
\code{\link{plot_semantic_viz}()},
\code{\link{plot_similarity_heatmap}()},
\code{\link{plot_term_trends_continuous}()},
\code{\link{plot_topic_effects_categorical}()},
\code{\link{plot_topic_effects_continuous}()},
\code{\link{plot_topic_probability}()},
\code{\link{plot_word_frequency}()},
\code{\link{plot_word_probability}()}
}
\concept{visualization}
