% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{get_spacy_embeddings}
\alias{get_spacy_embeddings}
\title{Get spaCy Word Embeddings}
\usage{
get_spacy_embeddings(texts, model = "en_core_web_md")
}
\arguments{
\item{texts}{Character vector of words or texts.}

\item{model}{Character; spaCy model to use (default: "en_core_web_md").}
}
\value{
A matrix of word embeddings (rows = texts, cols = dimensions).
}
\description{
Get word vector embeddings for words or texts using spaCy.
Requires a spaCy model with word vectors.
}
\seealso{
Other lexical:
\code{\link[=calculate_dispersion_metrics]{calculate_dispersion_metrics()}},
\code{\link[=calculate_lexical_dispersion]{calculate_lexical_dispersion()}},
\code{\link[=calculate_log_odds_ratio]{calculate_log_odds_ratio()}},
\code{\link[=calculate_text_readability]{calculate_text_readability()}},
\code{\link[=clear_lexdiv_cache]{clear_lexdiv_cache()}},
\code{\link[=detect_multi_words]{detect_multi_words()}},
\code{\link[=extract_keywords_keyness]{extract_keywords_keyness()}},
\code{\link[=extract_keywords_tfidf]{extract_keywords_tfidf()}},
\code{\link[=extract_morphology]{extract_morphology()}},
\code{\link[=extract_named_entities]{extract_named_entities()}},
\code{\link[=extract_noun_chunks]{extract_noun_chunks()}},
\code{\link[=extract_pos_tags]{extract_pos_tags()}},
\code{\link[=extract_subjects_objects]{extract_subjects_objects()}},
\code{\link[=find_similar_words]{find_similar_words()}},
\code{\link[=get_sentences]{get_sentences()}},
\code{\link[=get_spacy_model_info]{get_spacy_model_info()}},
\code{\link[=get_word_similarity]{get_word_similarity()}},
\code{\link[=init_spacy_nlp]{init_spacy_nlp()}},
\code{\link{lexical_analysis}},
\code{\link[=lexical_diversity_analysis]{lexical_diversity_analysis()}},
\code{\link[=lexical_frequency_analysis]{lexical_frequency_analysis()}},
\code{\link[=parse_morphology_string]{parse_morphology_string()}},
\code{\link[=plot_keyness_keywords]{plot_keyness_keywords()}},
\code{\link[=plot_keyword_comparison]{plot_keyword_comparison()}},
\code{\link[=plot_lexical_diversity_distribution]{plot_lexical_diversity_distribution()}},
\code{\link[=plot_morphology_feature]{plot_morphology_feature()}},
\code{\link[=plot_readability_by_group]{plot_readability_by_group()}},
\code{\link[=plot_readability_distribution]{plot_readability_distribution()}},
\code{\link[=plot_tfidf_keywords]{plot_tfidf_keywords()}},
\code{\link[=plot_top_readability_documents]{plot_top_readability_documents()}},
\code{\link[=render_displacy_dep]{render_displacy_dep()}},
\code{\link[=render_displacy_ent]{render_displacy_ent()}},
\code{\link[=spacy_extract_entities]{spacy_extract_entities()}},
\code{\link[=spacy_has_vectors]{spacy_has_vectors()}},
\code{\link[=spacy_initialized]{spacy_initialized()}},
\code{\link[=spacy_lemmatize]{spacy_lemmatize()}},
\code{\link[=spacy_parse_full]{spacy_parse_full()}},
\code{\link[=summarize_morphology]{summarize_morphology()}}
}
\concept{lexical}
