% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{get_topic_terms}
\alias{get_topic_terms}
\title{Select Top Terms for Each Topic}
\usage{
get_topic_terms(stm_model, top_term_n = 10, verbose = TRUE, ...)
}
\arguments{
\item{stm_model}{An STM model object.}

\item{top_term_n}{The number of top terms to display for each topic (default: 10).}

\item{verbose}{Logical, if TRUE, prints progress messages.}

\item{...}{Further arguments passed to tidytext::tidy.}
}
\value{
A data frame containing the top terms for each topic.
}
\description{
This function selects the top terms for each topic based on their word
probability distribution (beta).
}
\examples{
if (interactive()) {
  mydata <- TextAnalysisR::SpecialEduTech

  united_tbl <- TextAnalysisR::unite_cols(
    mydata,
    listed_vars = c("title", "keyword", "abstract")
  )

  tokens <- TextAnalysisR::prep_texts(united_tbl, text_field = "united_texts")

  dfm_object <- quanteda::dfm(tokens)

  out <- quanteda::convert(dfm_object, to = "stm")

stm_15 <- stm::stm(
  data = out$meta,
  documents = out$documents,
  vocab = out$vocab,
  max.em.its = 75,
  init.type = "Spectral",
  K = 15,
  prevalence = ~ reference_type + s(year),
  verbose = TRUE)

top_topic_terms <- TextAnalysisR::get_topic_terms(
  stm_model = stm_15,
  top_term_n = 10,
  verbose = TRUE
  )
print(top_topic_terms)
}
}
\seealso{
Other topic-modeling:
\code{\link[=analyze_semantic_evolution]{analyze_semantic_evolution()}},
\code{\link[=assess_embedding_stability]{assess_embedding_stability()}},
\code{\link[=auto_tune_embedding_topics]{auto_tune_embedding_topics()}},
\code{\link[=calculate_assignment_consistency]{calculate_assignment_consistency()}},
\code{\link[=calculate_keyword_stability]{calculate_keyword_stability()}},
\code{\link[=calculate_npmi]{calculate_npmi()}},
\code{\link[=calculate_semantic_drift]{calculate_semantic_drift()}},
\code{\link[=calculate_topic_diversity]{calculate_topic_diversity()}},
\code{\link[=calculate_topic_probability]{calculate_topic_probability()}},
\code{\link[=calculate_topic_stability]{calculate_topic_stability()}},
\code{\link[=extract_topic_terms_df]{extract_topic_terms_df()}},
\code{\link[=find_optimal_k]{find_optimal_k()}},
\code{\link[=find_topic_matches]{find_topic_matches()}},
\code{\link[=fit_embedding_model]{fit_embedding_model()}},
\code{\link[=fit_temporal_model]{fit_temporal_model()}},
\code{\link[=fit_topic_prevalence_model]{fit_topic_prevalence_model()}},
\code{\link[=generate_topic_labels]{generate_topic_labels()}},
\code{\link[=get_topic_prevalence]{get_topic_prevalence()}},
\code{\link[=get_topic_texts]{get_topic_texts()}},
\code{\link[=identify_topic_trends]{identify_topic_trends()}},
\code{\link[=plot_model_comparison]{plot_model_comparison()}},
\code{\link[=plot_quality_metrics]{plot_quality_metrics()}},
\code{\link[=plot_topic_effects_categorical]{plot_topic_effects_categorical()}},
\code{\link[=plot_topic_effects_continuous]{plot_topic_effects_continuous()}},
\code{\link[=plot_topic_probability]{plot_topic_probability()}},
\code{\link[=plot_word_probability]{plot_word_probability()}},
\code{\link[=run_neural_topics_internal]{run_neural_topics_internal()}},
\code{\link[=validate_semantic_coherence]{validate_semantic_coherence()}}
}
\concept{topic-modeling}
