% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{get_topic_terms}
\alias{get_topic_terms}
\title{Select Top Terms for Each Topic}
\usage{
get_topic_terms(stm_model, top_term_n = 10, verbose = TRUE, ...)
}
\arguments{
\item{stm_model}{An STM model object.}

\item{top_term_n}{The number of top terms to display for each topic (default: 10).}

\item{verbose}{Logical, if TRUE, prints progress messages.}

\item{...}{Further arguments passed to tidytext::tidy.}
}
\value{
A data frame containing the top terms for each topic.
}
\description{
This function selects the top terms for each topic based on their word
probability distribution (beta).
}
\examples{
if (interactive()) {
  mydata <- TextAnalysisR::SpecialEduTech

  united_tbl <- TextAnalysisR::unite_cols(
    mydata,
    listed_vars = c("title", "keyword", "abstract")
  )

  tokens <- TextAnalysisR::prep_texts(united_tbl, text_field = "united_texts")

  dfm_object <- quanteda::dfm(tokens)

  stm_15 <- TextAnalysisR::create_stm_model(
  dfm_object,
  topic_n = 15,
  max.em.its = 75,
  categorical_var = "reference_type",
  continuous_var = "year",
  verbose = TRUE
  )

  out <- quanteda::convert(dfm_object, to = "stm")

stm_15 <- stm::stm(
  data = out$meta,
  documents = out$documents,
  vocab = out$vocab,
  max.em.its = 75,
  init.type = "Spectral",
  K = 15,
  prevalence = ~ reference_type + s(year),
  verbose = TRUE)

top_topic_terms <- TextAnalysisR::get_topic_terms(
  stm_model = stm_15,
  top_term_n = 10,
  verbose = TRUE
  )
print(top_topic_terms)
}
}
\seealso{
Other topic-modeling: 
\code{\link{analyze_semantic_evolution}()},
\code{\link{assess_embedding_stability}()},
\code{\link{assess_hybrid_stability}()},
\code{\link{auto_tune_embedding_topics}()},
\code{\link{calculate_assignment_consistency}()},
\code{\link{calculate_eval_metrics_internal}()},
\code{\link{calculate_keyword_stability}()},
\code{\link{calculate_semantic_drift}()},
\code{\link{calculate_topic_probability}()},
\code{\link{calculate_topic_stability}()},
\code{\link{find_optimal_k}()},
\code{\link{find_topic_matches}()},
\code{\link{fit_embedding_model}()},
\code{\link{fit_hybrid_model}()},
\code{\link{fit_temporal_model}()},
\code{\link{generate_topic_labels}()},
\code{\link{get_topic_prevalence}()},
\code{\link{get_topic_texts}()},
\code{\link{identify_topic_trends}()},
\code{\link{plot_model_comparison}()},
\code{\link{plot_quality_metrics}()},
\code{\link{run_contrastive_topics_internal}()},
\code{\link{run_neural_topics_internal}()},
\code{\link{run_temporal_topics_internal}()},
\code{\link{validate_semantic_coherence}()}
}
\concept{topic-modeling}
