% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{lexical_analysis}
\alias{lexical_analysis}
\title{Lexical Analysis Functions}
\description{
Comprehensive functions for lexical analysis including:
\itemize{
\item Linguistic Annotation (POS tagging, NER)
\item Frequency Analysis (word frequency, n-grams, MWEs)
\item Keywords (TF-IDF, keyness)
\item Lexical Diversity (TTR, MTLD, MATTR)
\item Readability (Flesch, Gunning Fog, etc.)
}
}
\seealso{
Other lexical: 
\code{\link{calculate_dispersion_metrics}()},
\code{\link{calculate_lexical_dispersion}()},
\code{\link{calculate_log_odds_ratio}()},
\code{\link{calculate_text_readability}()},
\code{\link{clear_lexdiv_cache}()},
\code{\link{detect_multi_words}()},
\code{\link{extract_keywords_keyness}()},
\code{\link{extract_keywords_tfidf}()},
\code{\link{extract_morphology}()},
\code{\link{extract_named_entities}()},
\code{\link{extract_noun_chunks}()},
\code{\link{extract_pos_tags}()},
\code{\link{extract_subjects_objects}()},
\code{\link{find_similar_words}()},
\code{\link{get_sentences}()},
\code{\link{get_spacy_embeddings}()},
\code{\link{get_spacy_model_info}()},
\code{\link{get_word_similarity}()},
\code{\link{init_spacy_nlp}()},
\code{\link{lexical_diversity_analysis}()},
\code{\link{lexical_frequency_analysis}()},
\code{\link{parse_morphology_string}()},
\code{\link{plot_keyness_keywords}()},
\code{\link{plot_keyword_comparison}()},
\code{\link{plot_lexical_diversity_distribution}()},
\code{\link{plot_morphology_feature}()},
\code{\link{plot_readability_by_group}()},
\code{\link{plot_readability_distribution}()},
\code{\link{plot_tfidf_keywords}()},
\code{\link{plot_top_readability_documents}()},
\code{\link{render_displacy_dep}()},
\code{\link{render_displacy_ent}()},
\code{\link{spacy_extract_entities}()},
\code{\link{spacy_has_vectors}()},
\code{\link{spacy_initialized}()},
\code{\link{spacy_parse_full}()},
\code{\link{summarize_morphology}()}
}
\concept{lexical}
