% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readability_analysis.R
\name{lexical_diversity_analysis}
\alias{lexical_diversity_analysis}
\title{Lexical Diversity Analysis}
\usage{
lexical_diversity_analysis(dfm_object, measures = "all")
}
\arguments{
\item{dfm_object}{A document-feature matrix from quanteda}

\item{measures}{Character vector of measures to calculate. Options:
"all", "MTLD" (recommended), "MATTR" (recommended), "MSTTR", "TTR", "CTTR", "Maas", "K", "D"}
}
\value{
A list with lexical_diversity (data frame) and summary_stats
}
\description{
Calculates lexical diversity metrics to measure vocabulary richness.
MTLD and MATTR are most stable and text-length independent.
}
\examples{
\dontrun{
data(SpecialEduTech)
texts <- SpecialEduTech$abstract[1:10]
corp <- quanteda::corpus(texts)
toks <- quanteda::tokens(corp)
dfm_obj <- quanteda::dfm(toks)
lex_div <- lexical_diversity_analysis(dfm_obj)
print(lex_div)
}

}
