% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readability_analysis.R
\name{lexical_frequency_analysis}
\alias{lexical_frequency_analysis}
\title{Lexical Frequency Analysis}
\usage{
lexical_frequency_analysis(...)
}
\arguments{
\item{...}{Arguments passed to plot_word_frequency}
}
\description{
Wrapper function for plot_word_frequency for lexical analysis.
}
