% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{lexical_frequency_analysis}
\alias{lexical_frequency_analysis}
\title{Lexical Frequency Analysis}
\usage{
lexical_frequency_analysis(...)
}
\arguments{
\item{...}{Arguments passed to plot_word_frequency}
}
\description{
Wrapper function for plot_word_frequency for lexical analysis.
}
\seealso{
Other lexical: 
\code{\link{calculate_text_readability}()},
\code{\link{clear_lexdiv_cache}()},
\code{\link{detect_multi_words}()},
\code{\link{extract_keywords_keyness}()},
\code{\link{extract_keywords_tfidf}()},
\code{\link{extract_named_entities}()},
\code{\link{extract_pos_tags}()},
\code{\link{lexical_analysis}},
\code{\link{lexical_diversity_analysis}()},
\code{\link{plot_keyness_keywords}()},
\code{\link{plot_keyword_comparison}()},
\code{\link{plot_lexical_diversity_distribution}()},
\code{\link{plot_readability_by_group}()},
\code{\link{plot_readability_distribution}()},
\code{\link{plot_tfidf_keywords}()},
\code{\link{plot_top_readability_documents}()}
}
\concept{lexical}
