% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{plot_keyword_comparison}
\alias{plot_keyword_comparison}
\title{Plot Keyword Comparison (TF-IDF vs Frequency)}
\usage{
plot_keyword_comparison(
  tfidf_data,
  top_n = 10,
  title = NULL,
  normalized = FALSE
)
}
\arguments{
\item{tfidf_data}{Data frame from extract_keywords_tfidf()}

\item{top_n}{Number of keywords to display (default: 10)}

\item{title}{Plot title (default: auto-generated)}

\item{normalized}{Logical, whether TF-IDF scores are normalized (default: FALSE)}
}
\value{
A plotly grouped bar chart
}
\description{
Creates a grouped bar plot comparing TF-IDF scores with term frequencies.
}
\seealso{
Other lexical:
\code{\link[=calculate_dispersion_metrics]{calculate_dispersion_metrics()}},
\code{\link[=calculate_lexical_dispersion]{calculate_lexical_dispersion()}},
\code{\link[=calculate_log_odds_ratio]{calculate_log_odds_ratio()}},
\code{\link[=calculate_text_readability]{calculate_text_readability()}},
\code{\link[=clear_lexdiv_cache]{clear_lexdiv_cache()}},
\code{\link[=detect_multi_words]{detect_multi_words()}},
\code{\link[=extract_keywords_keyness]{extract_keywords_keyness()}},
\code{\link[=extract_keywords_tfidf]{extract_keywords_tfidf()}},
\code{\link[=extract_morphology]{extract_morphology()}},
\code{\link[=extract_named_entities]{extract_named_entities()}},
\code{\link[=extract_noun_chunks]{extract_noun_chunks()}},
\code{\link[=extract_pos_tags]{extract_pos_tags()}},
\code{\link[=extract_subjects_objects]{extract_subjects_objects()}},
\code{\link[=find_similar_words]{find_similar_words()}},
\code{\link[=get_sentences]{get_sentences()}},
\code{\link[=get_spacy_embeddings]{get_spacy_embeddings()}},
\code{\link[=get_spacy_model_info]{get_spacy_model_info()}},
\code{\link[=get_word_similarity]{get_word_similarity()}},
\code{\link[=init_spacy_nlp]{init_spacy_nlp()}},
\code{\link{lexical_analysis}},
\code{\link[=lexical_diversity_analysis]{lexical_diversity_analysis()}},
\code{\link[=lexical_frequency_analysis]{lexical_frequency_analysis()}},
\code{\link[=parse_morphology_string]{parse_morphology_string()}},
\code{\link[=plot_keyness_keywords]{plot_keyness_keywords()}},
\code{\link[=plot_lexical_diversity_distribution]{plot_lexical_diversity_distribution()}},
\code{\link[=plot_morphology_feature]{plot_morphology_feature()}},
\code{\link[=plot_readability_by_group]{plot_readability_by_group()}},
\code{\link[=plot_readability_distribution]{plot_readability_distribution()}},
\code{\link[=plot_tfidf_keywords]{plot_tfidf_keywords()}},
\code{\link[=plot_top_readability_documents]{plot_top_readability_documents()}},
\code{\link[=render_displacy_dep]{render_displacy_dep()}},
\code{\link[=render_displacy_ent]{render_displacy_ent()}},
\code{\link[=spacy_extract_entities]{spacy_extract_entities()}},
\code{\link[=spacy_has_vectors]{spacy_has_vectors()}},
\code{\link[=spacy_initialized]{spacy_initialized()}},
\code{\link[=spacy_lemmatize]{spacy_lemmatize()}},
\code{\link[=spacy_parse_full]{spacy_parse_full()}},
\code{\link[=summarize_morphology]{summarize_morphology()}}
}
\concept{lexical}
