% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{plot_tfidf_keywords}
\alias{plot_tfidf_keywords}
\title{Plot TF-IDF Keywords}
\usage{
plot_tfidf_keywords(tfidf_data, title = NULL, normalized = FALSE)
}
\arguments{
\item{tfidf_data}{Data frame from extract_keywords_tfidf()}

\item{title}{Plot title (default: "Top Keywords by TF-IDF Score")}

\item{normalized}{Logical, whether scores are normalized (for label) (default: FALSE)}
}
\value{
A plotly bar chart
}
\description{
Creates a horizontal bar plot of top keywords by TF-IDF score.
}
\seealso{
Other lexical: 
\code{\link{calculate_text_readability}()},
\code{\link{clear_lexdiv_cache}()},
\code{\link{detect_multi_words}()},
\code{\link{extract_keywords_keyness}()},
\code{\link{extract_keywords_tfidf}()},
\code{\link{extract_named_entities}()},
\code{\link{extract_pos_tags}()},
\code{\link{lexical_analysis}},
\code{\link{lexical_diversity_analysis}()},
\code{\link{lexical_frequency_analysis}()},
\code{\link{plot_keyness_keywords}()},
\code{\link{plot_keyword_comparison}()},
\code{\link{plot_lexical_diversity_distribution}()},
\code{\link{plot_readability_by_group}()},
\code{\link{plot_readability_distribution}()},
\code{\link{plot_top_readability_documents}()}
}
\concept{lexical}
