% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_topic_term}
\alias{plot_topic_term}
\title{Visualize a plot for highest word probabilities within each topic}
\usage{
plot_topic_term(data, top_n, topic_names = NULL, ...)
}
\arguments{
\item{data}{A tidy data frame that includes term (word)-topic probabilities
(probabilities of each word per each topic).}

\item{top_n}{A number of top n terms frequently observed in each document.}

\item{topic_names}{(Labeled) topic names}

\item{...}{Further arguments passed to \code{dplyr::group_by}.}
}
\value{
A ggplot object output from \code{stm::stm}, \code{tidytext::tidy}, and \code{ggplot2::ggplot}.
The result is a ggplot object representing the topic-term plot.
}
\description{
Visualize a plot for highest word probabilities within each topic.
}
\examples{
if(requireNamespace("quanteda", "tidytext")){
dfm <- SpecialEduTech \%>\% preprocess_texts(text_field = "abstract") \%>\% quanteda::dfm()
data <- tidytext::tidy(stm_15, document_names = rownames(dfm), log = FALSE)
data \%>\% plot_topic_term(top_n = 10)
}

}
