% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_topic_term}
\alias{plot_topic_term}
\title{Plot topic per-term per-topic probabilities}
\usage{
plot_topic_term(data, top_n, ncol = 3, topic_names = NULL, ...)
}
\arguments{
\item{data}{A tidy data frame that includes per-term per-topic probabilities (beta).}

\item{top_n}{A number of highest per-term per-topic probabilities in each document (number of top_n can be changed).}

\item{ncol}{A number of columns in the facet plot.}

\item{topic_names}{(Labeled) topic names}

\item{...}{Further arguments passed to \code{dplyr::group_by}.}
}
\value{
A ggplot object output from \code{stm::stm}, \code{tidytext::tidy}, and \code{ggplot2::ggplot}.
The result is a ggplot object representing the topic-term plot.
}
\description{
Plot per-term per-topic probabilities with highest word probabilities.
}
\examples{
suppressWarnings({
if(requireNamespace("quanteda", "tidytext")){
dfm <- SpecialEduTech \%>\%
       preprocess_texts(text_field = "abstract") \%>\%
       quanteda::dfm()
data <- tidytext::tidy(stm_15, document_names = rownames(dfm), log = FALSE)
data \%>\% plot_topic_term(top_n = 2, ncol = 3)
}
})

}
