% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_word_co_occurrence_network}
\alias{plot_word_co_occurrence_network}
\title{Plot a Word Co-occurrence Network}
\usage{
plot_word_co_occurrence_network(
  dfm_object,
  term_col = "term",
  document = "document",
  co_occur_n = 2,
  height = 700,
  width = 1000
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{term_col}{The column name for terms (default is "term").}

\item{document}{The column name for document IDs (default is "document").}

\item{co_occur_n}{Minimum number of co-occurrences for filtering terms (default is 2).}

\item{height}{The height of the resulting Plotly plot, in pixels. Defaults to \code{700}.}

\item{width}{The width of the resulting Plotly plot, in pixels. Defaults to \code{1000}.}
}
\value{
A Plotly object visualizing the interactive word co-occurrence network.
}
\description{
Visualize the co-occurrence relationships between terms in the corpus based on pairwise counts.
}
\examples{
if (interactive()) {
 df <- TextAnalysisR::SpecialEduTech
 dfm_object <- df \%>\%
   TextAnalysisR::preprocess_texts(text_field = "abstract") \%>\%
   quanteda::dfm() \%>\%
   quanteda::dfm_trim(min_termfreq = 1, min_docfreq = 1) \%>\%
   .[quanteda::ntoken(.) > 0, ]
TextAnalysisR::plot_word_co_occurrence_network(dfm_object,
                               term_col = "abstract",
                               document = "document",
                               co_occur_n = 5,
                               height = 700,
                               width = 1000)
}


}
