% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_word_co_occurrence_network}
\alias{plot_word_co_occurrence_network}
\title{Plot a Word Co-occurrence Network}
\usage{
plot_word_co_occurrence_network(
  dfm_td,
  term_col = "term",
  doc_col = "document",
  co_occur_n = 2
)
}
\arguments{
\item{dfm_td}{A tidy data frame of terms (output of \code{tidytext::tidy(dfm_object)}),
where \code{dfm_object} is a document-feature matrix created using \code{quanteda::dfm()}
after preprocessing text data.}

\item{term_col}{The column name for terms (default is "term").}

\item{doc_col}{The column name for document IDs (default is "document").}

\item{co_occur_n}{Minimum number of co-occurrences for filtering terms (default is 2).}
}
\value{
A \code{ggraph} object visualizing the word co-occurrence network.
}
\description{
Visualize the co-occurrence relationships between terms in the corpus based on pairwise counts.
}
\examples{
if (interactive()) {
df <- TextAnalysisR::SpecialEduTech
dfm_object <- preprocess_texts(df, text_field = "text") \%>\%
 quanteda::dfm()
dfm_td <- tidytext::tidy(dfm_object)
plot_word_co_occurrence_network(dfm_td,
                               term_col = "abstract",
                               doc_col = "document",
                               co_occur_n = 5)
}


}
