% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_word_correlation_network}
\alias{plot_word_correlation_network}
\title{Plot a Word Correlation Network}
\usage{
plot_word_correlation_network(
  dfm_td,
  term_col = "term",
  doc_col = "document",
  co_occur_n = 2,
  correlation_threshold = 0.3
)
}
\arguments{
\item{dfm_td}{A tidy data frame of terms (output of \code{tidytext::tidy(dfm_object)}),
where \code{dfm_object} is a document-feature matrix created using \code{quanteda::dfm()}
after preprocessing text data.}

\item{term_col}{The column name for terms (default is "term").}

\item{doc_col}{The column name for document IDs (default is "document").}

\item{co_occur_n}{Minimum number of co-occurrences for filtering terms (default is 2).}

\item{correlation_threshold}{Minimum correlation value to include edges in the graph (default is 0.3).}
}
\value{
A \code{ggraph} object visualizing the word correlation network.
}
\description{
Visualize the correlation relationships between terms in the corpus based on pairwise correlations.
}
\examples{
if (interactive()) {
df <- TextAnalysisR::SpecialEduTech
dfm_object <- preprocess_texts(df, text_field = "text") \%>\%
 quanteda::dfm()
dfm_td <- tidytext::tidy(dfm_object)
plot_word_correlation_network(dfm_td,
                             term_col = "abstract",
                             doc_col = "document",
                             correlation_threshold = 0.3)
}


}
