% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_word_frequency}
\alias{plot_word_frequency}
\title{Plot word frequency results.}
\usage{
plot_word_frequency(data, n = 20, ...)
}
\arguments{
\item{data}{A document-feature matrix (dfm) object through the quanteda package.}

\item{n}{The number of top n features (terms or words).}

\item{...}{Further arguments passed to \code{quanteda.textstats::textstat_frequency}.}
}
\value{
A ggplot object output from \code{quanteda.textstats::textstat_frequency} and \code{ggplot2::ggplot}.
The result is a ggplot object representing the word frequency plot.
}
\description{
Plot the frequently observed top n terms.
}
\examples{
if(requireNamespace("quanteda")){
dfm <- SpecialEduTech \%>\%
       preprocess_texts(text_field = "abstract",
       verbose = FALSE) \%>\%
       quanteda::dfm()
dfm \%>\% plot_word_frequency(n = 20)
}

}
