% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_word_frequency}
\alias{plot_word_frequency}
\title{Plot Word Frequency}
\usage{
plot_word_frequency(dfm_object, n = 20, ...)
}
\arguments{
\item{dfm_object}{A \code{quanteda} dfm object.}

\item{n}{The number of top features (terms or words) to display.}

\item{...}{Further arguments passed to \code{quanteda.textstats::textstat_frequency}.}
}
\value{
A \code{ggplot} object visualizing the top terms by their frequency. The plot
shows each term on one axis and frequency on the other, with points representing their
observed frequencies.
}
\description{
Given a document-feature matrix (dfm), this function computes the most frequent terms
and creates a ggplot-based visualization of term frequencies.
}
\examples{
if (interactive()) {
df <- TextAnalysisR::SpecialEduTech
dfm_object <- df \%>\%
  TextAnalysisR::preprocess_texts(text_field = "abstract") \%>\%
  quanteda::dfm() \%>\%
  quanteda::dfm_trim(min_termfreq = 1, min_docfreq = 1) \%>\%
  .[quanteda::ntoken(.) > 0, ]
TextAnalysisR::plot_word_frequency(dfm_object, n = 20)
}
}
