% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{plot_word_probabilities}
\alias{plot_word_probabilities}
\title{Plot Highest Word Probabilities for Each Topic}
\usage{
plot_word_probabilities(
  dfm_object,
  topic_n,
  max.em.its = 75,
  categorical_var = NULL,
  continuous_var = NULL,
  top_term_n = 10,
  ncol = 3,
  topic_names = NULL,
  height = 1200,
  width = 800,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{topic_n}{The number of topics to display.}

\item{max.em.its}{Maximum number of EM iterations (default: 75).}

\item{categorical_var}{An optional character string for a categorical variable in the metadata.}

\item{continuous_var}{An optional character string for a continuous variable in the metadata.}

\item{top_term_n}{The number of top terms to display for each topic (default: 10).}

\item{ncol}{The number of columns in the facet plot (default: 3).}

\item{topic_names}{An optional character vector for labeling topics. If provided, must be the same length as the number of topics.}

\item{height}{The height of the resulting Plotly plot, in pixels. Defaults to \code{1200}.}

\item{width}{The width of the resulting Plotly plot, in pixels. Defaults to \code{800}.}

\item{verbose}{Logical; if \code{TRUE}, prints progress information.}

\item{...}{Further arguments passed to \code{stm::searchK}.}
}
\value{
A \code{Plotly} object showing a facet-wrapped chart of top terms for each topic,
ordered by their per-topic probability (beta). Each facet represents a topic.
}
\description{
This function provides a visualization of the top terms for each topic,
ordered by their word probability distribution for each topic (beta).
}
\details{
If \code{topic_names} is provided, it replaces the default "Topic \{n\}" labels with custom names.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
TextAnalysisR::plot_word_probabilities(
  dfm_object = dfm_object,
  topic_n = 15,
  max.em.its = 75,
  categorical_var = "reference_type",
  continuous_var = "year",
  top_term_n = 10,
  ncol = 3,
  height = 1200,
  width = 800,
  verbose = TRUE)
}

}
