% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{preprocess_texts}
\alias{preprocess_texts}
\title{Preprocess Text Data}
\usage{
preprocess_texts(
  df,
  text_field = "united_texts",
  custom_stopwords = NULL,
  min_char = 2,
  ...
)
}
\arguments{
\item{df}{A data frame that contains text data.}

\item{text_field}{The name of the column containing text data.}

\item{custom_stopwords}{A character vector of additional stopwords to remove. Default is NULL.}

\item{min_char}{Minimum length in characters for tokens (default is 2).}

\item{...}{Further arguments passed to \code{quanteda::corpus}.}
}
\value{
A \code{quanteda::tokens} object. The output is a list of tokens. Each token is a word in the text data.
The function removes punctuation, symbols, numbers, URLs, and stopwords. It also converts all words to lowercase.
}
\description{
Preprocesses text data by:
\itemize{
\item Constructing a corpus
\item Tokenizing text into words
\item Converting to lowercase
\item Removing default English stopwords and optional custom stopwords
\item Specifying a minimum token length.
}

Typically used before constructing a dfm and fitting an STM model.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  tokens <- preprocess_texts(df, text_field = "abstract")
  tokens
}
}
