% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{preprocess_texts}
\alias{preprocess_texts}
\title{Preprocess Text Data}
\usage{
preprocess_texts(
  united_tbl,
  text_field = "united_texts",
  custom_stopwords = NULL,
  min_char = 2,
  ...
)
}
\arguments{
\item{united_tbl}{A data frame that contains text data.}

\item{text_field}{The name of the column in \code{united_tbl} that contains text data.}

\item{custom_stopwords}{A character vector of additional stopwords to remove. Default is NULL.}

\item{min_char}{Minimum length in characters for tokens (default is 2).}

\item{...}{Further arguments passed to \code{quanteda::corpus}.}
}
\value{
A \code{quanteda} tokens object. Each row in the object represents a document, and each column represents a token.
The object is ready for constructing a dfm and fitting an STM model.
}
\description{
Preprocesses text data by:
\itemize{
\item Constructing a corpus
\item Tokenizing text into words
\item Converting to lowercase
\item Removing default English stopwords and optional custom stopwords
\item Specifying a minimum token length.
}

Typically used before constructing a dfm and fitting an STM model.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  tokens
}
}
