% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_notification_by_id}
\alias{remove_notification_by_id}
\title{Remove Notification by ID}
\usage{
remove_notification_by_id(id)
}
\arguments{
\item{id}{The notification ID to remove}
}
\value{
Removes a Shiny notification. Returns NULL invisibly.
}
\description{
Removes a notification with a specific ID.
}
