% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch and browse the TextAnalysisR app}
\usage{
run_app()
}
\value{
No return value, called for side effects (launching Shiny app)
}
\description{
Launch and browse the TextAnalysisR app.
}
\examples{
if (interactive()) {
  library(TextAnalysisR)
  run_app()
}

}
