% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_text_input}
\alias{sanitize_text_input}
\title{Sanitize Text Input}
\usage{
sanitize_text_input(text)
}
\arguments{
\item{text}{Text input from user}
}
\value{
Sanitized text
}
\description{
Sanitize Text Input
}
\keyword{internal}
