% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_analysis.R
\name{semantic_similarity_analysis}
\alias{semantic_similarity_analysis}
\title{Semantic Similarity Analysis}
\usage{
semantic_similarity_analysis(...)
}
\arguments{
\item{...}{Arguments passed to calculate_document_similarity}
}
\value{
Similarity analysis results
}
\description{
Wrapper for calculate_document_similarity
}
