% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_guide_modal}
\alias{show_guide_modal}
\title{Show Guide Modal Dialog from HTML File}
\usage{
show_guide_modal(guide_name, title, size = "l")
}
\arguments{
\item{guide_name}{Name of the guide file (without .html extension).
Files should be located in inst/TextAnalysisR.app/markdown/guides/}

\item{title}{Modal dialog title to display}

\item{size}{Size of the modal dialog (default: "l" for large).
Options: "s" (small), "m" (medium), "l" (large)}
}
\value{
Displays a Shiny modal dialog. Returns NULL invisibly.
}
\description{
Loads and displays a modal dialog with guide content from an HTML file.
This function is designed for Shiny applications to display help documentation
stored in external HTML files, reducing server.R file size and improving
maintainability.
}
\details{
Guide HTML files should be placed in:
\code{inst/TextAnalysisR.app/markdown/guides/<guide_name>.html}

The function will look for the guide file in the installed package location.
If the file is not found, it displays an error message in the modal.
}
\examples{
\dontrun{
observeEvent(input$showDimRedInfo, {
  show_guide_modal("dimensionality_reduction_guide", "Dimensionality Reduction Guide")
})

observeEvent(input$showClusteringInfo, {
  show_guide_modal("clustering_guide", "Document Clustering Guide")
})
}

}
