% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{spacy_has_vectors}
\alias{spacy_has_vectors}
\title{Check if Model Has Word Vectors}
\usage{
spacy_has_vectors()
}
\value{
Logical; TRUE if model has vectors, FALSE otherwise.
}
\description{
Check if the loaded spaCy model has word vectors for similarity calculations.
}
\concept{lexical}
