% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_analysis.R
\name{spacy_initialized}
\alias{spacy_initialized}
\title{Check if spaCy is Initialized}
\usage{
spacy_initialized()
}
\value{
Logical; TRUE if initialized, FALSE otherwise.
}
\description{
Check whether spaCy has been initialized.
}
\concept{lexical}
