% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysisR-package.R
\docType{package}
\name{TextAnalysisR-package}
\alias{TextAnalysisR}
\alias{TextAnalysisR-package}
\title{TextAnalysisR: A text mining workflow tool}
\description{
'TextAnalysisR' provides a supporting workflow tool for text mining analysis. The web app incorporates 'quanteda' (text preprocessing), 'stm' (structural topic modeling), 'ggraph', and 'widyr' (network analysis). 'tidytext' is implemented to tidy non-tidy format objects. The R Shiny web app is available at 'TextAnalysisR::TextAnalysisR.app()' or 'https://textanalysisr.org'. Functions are provided for completing word-topic probabilities, document-topic probabilities, estimated effects of covariates on topic prevalence, and network analysis, similar to tasks available in the web app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mshin77.github.io/TextAnalysisR}
  \item Report bugs at \url{https://github.com/mshin77/TextAnalysisR/issues}
}

}
\author{
\strong{Maintainer}: Mikyung Shin \email{shin.mikyung@gmail.com} (\href{https://orcid.org/0000-0001-7907-9193}{ORCID})

}
\keyword{internal}
