% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysisR-package.R
\docType{package}
\name{TextAnalysisR-package}
\alias{TextAnalysisR}
\alias{TextAnalysisR-package}
\title{TextAnalysisR: Text Mining Workflow Tools}
\description{
'TextAnalysisR' provides a supporting workflow for text mining analysis. The web app incorporates 'quanteda' (text preprocessing), 'stm' (structural topic modeling), and 'ggraph' as well as 'widyr' (network analysis). 'tidytext' was implemented to tidy non-tidy format objects. These steps are similar to those demonstrated in the Shiny web app at 'TextAnalysisR::TextAnalysisR.app()'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mshin77.github.io/TextAnalysisR/}
  \item Report bugs at \url{https://github.com/mshin77/TextAnalysisR/issues}
}

}
\author{
\strong{Maintainer}: Mikyung Shin \email{shin.mikyung@gmail.com} (\href{https://orcid.org/0000-0001-7907-9193}{ORCID})

}
\keyword{internal}
