% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{topic_probability_table}
\alias{topic_probability_table}
\title{Create a Table for Per-Document Per-Topic Probabilities}
\usage{
topic_probability_table(stm_model, top_n = 10, verbose = TRUE, ...)
}
\arguments{
\item{stm_model}{A fitted STM model object.}

\item{top_n}{The number of topics to display, ordered by their mean prevalence.}

\item{verbose}{Logical, if TRUE, prints progress messages.}

\item{...}{Further arguments passed to \code{tidytext::tidy}.}
}
\value{
A \code{tibble} containing columns \code{topic} and \code{gamma}, where \code{topic}
is a factor representing each topic (relabeled with a "Topic X" format), and \code{gamma} is the
mean topic prevalence across all documents. Numeric values are rounded to three decimal places.
}
\description{
This function generates a table of mean topic prevalence across all documents.
}
\examples{
if (interactive()) {

df <- TextAnalysisR::SpecialEduTech

 united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))

 tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")

 dfm_object <- quanteda::dfm(tokens)

 out <- quanteda::convert(dfm_object, to = "stm")

stm_15 <- stm::stm(
  data = out$meta,
  documents = out$documents,
  vocab = out$vocab,
  max.em.its = 75,
  init.type = "Spectral",
  K = 15,
  prevalence = ~ reference_type + s(year),
  verbose = TRUE)

topic_probability_table <- TextAnalysisR::topic_probability_table(
   stm_model= stm_15,
   top_n = 10,
   verbose = TRUE)

print(topic_probability_table)
}
}
