% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{topic_probability_table}
\alias{topic_probability_table}
\title{Visualize a table for per-document per-topic probabilities}
\usage{
topic_probability_table(data, top_n, ...)
}
\arguments{
\item{data}{A tidy data frame that includes per-document per-topic probabilities (gamma).}

\item{top_n}{A number of highest per-document per-topic probabilities (number of top_n can be changed).}

\item{...}{Further arguments passed.}
}
\value{
A tibble (data frame) object with a list of topic probabilities from \code{tidytext::tidy}.
The result is a ggplot object representing the topic-term plot.
}
\description{
Create a table of per-document per-topic probabilities.
}
\examples{
suppressWarnings({
if(requireNamespace("quanteda", "tidytext")){
dfm <- SpecialEduTech \%>\%
       preprocess_texts(text_field = "abstract") \%>\%
       quanteda::dfm()
data <- tidytext::tidy(stm_15, matrix = "gamma", document_names = rownames(dfm), log = FALSE)
data \%>\% topic_probability_table(top_n = 15)
}
})

}
