% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{validate_semantic_coherence}
\alias{validate_semantic_coherence}
\title{Validate Semantic Coherence}
\usage{
validate_semantic_coherence(embeddings, topic_assignments, ...)
}
\arguments{
\item{embeddings}{Document embeddings matrix.}

\item{topic_assignments}{Vector of topic assignments for documents.}

\item{...}{Additional parameters (currently unused).}
}
\value{
List containing coherence score and metrics.
}
\description{
Validates the semantic coherence of topic assignments using
intra-cluster distance in embedding space.
}
\seealso{
Other topic-modeling:
\code{\link[=analyze_semantic_evolution]{analyze_semantic_evolution()}},
\code{\link[=assess_embedding_stability]{assess_embedding_stability()}},
\code{\link[=assess_hybrid_stability]{assess_hybrid_stability()}},
\code{\link[=auto_tune_embedding_topics]{auto_tune_embedding_topics()}},
\code{\link[=calculate_assignment_consistency]{calculate_assignment_consistency()}},
\code{\link[=calculate_keyword_stability]{calculate_keyword_stability()}},
\code{\link[=calculate_npmi]{calculate_npmi()}},
\code{\link[=calculate_semantic_drift]{calculate_semantic_drift()}},
\code{\link[=calculate_topic_diversity]{calculate_topic_diversity()}},
\code{\link[=calculate_topic_probability]{calculate_topic_probability()}},
\code{\link[=calculate_topic_stability]{calculate_topic_stability()}},
\code{\link[=extract_topic_terms_df]{extract_topic_terms_df()}},
\code{\link[=find_optimal_k]{find_optimal_k()}},
\code{\link[=find_topic_matches]{find_topic_matches()}},
\code{\link[=fit_embedding_model]{fit_embedding_model()}},
\code{\link[=fit_hybrid_model]{fit_hybrid_model()}},
\code{\link[=fit_temporal_model]{fit_temporal_model()}},
\code{\link[=fit_topic_prevalence_model]{fit_topic_prevalence_model()}},
\code{\link[=generate_topic_labels]{generate_topic_labels()}},
\code{\link[=get_topic_prevalence]{get_topic_prevalence()}},
\code{\link[=get_topic_terms]{get_topic_terms()}},
\code{\link[=get_topic_texts]{get_topic_texts()}},
\code{\link[=identify_topic_trends]{identify_topic_trends()}},
\code{\link[=plot_model_comparison]{plot_model_comparison()}},
\code{\link[=plot_quality_metrics]{plot_quality_metrics()}},
\code{\link[=plot_topic_effects_categorical]{plot_topic_effects_categorical()}},
\code{\link[=plot_topic_effects_continuous]{plot_topic_effects_continuous()}},
\code{\link[=plot_topic_probability]{plot_topic_probability()}},
\code{\link[=plot_word_probability]{plot_word_probability()}},
\code{\link[=run_neural_topics_internal]{run_neural_topics_internal()}}
}
\concept{topic-modeling}
