% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{validate_semantic_coherence}
\alias{validate_semantic_coherence}
\title{Validate Semantic Coherence}
\usage{
validate_semantic_coherence(embeddings, topic_assignments, ...)
}
\arguments{
\item{embeddings}{Document embeddings matrix}

\item{topic_assignments}{Vector of topic assignments for documents}

\item{...}{Additional parameters}
}
\value{
List containing coherence score and metrics
}
\description{
Validates the semantic coherence of topic assignments
}
\seealso{
Other topic-modeling: 
\code{\link{analyze_semantic_evolution}()},
\code{\link{assess_embedding_stability}()},
\code{\link{assess_hybrid_stability}()},
\code{\link{auto_tune_embedding_topics}()},
\code{\link{calculate_assignment_consistency}()},
\code{\link{calculate_eval_metrics_internal}()},
\code{\link{calculate_keyword_stability}()},
\code{\link{calculate_semantic_drift}()},
\code{\link{calculate_topic_probability}()},
\code{\link{calculate_topic_stability}()},
\code{\link{find_optimal_k}()},
\code{\link{find_topic_matches}()},
\code{\link{fit_embedding_model}()},
\code{\link{fit_hybrid_model}()},
\code{\link{fit_temporal_model}()},
\code{\link{generate_topic_labels}()},
\code{\link{get_topic_prevalence}()},
\code{\link{get_topic_terms}()},
\code{\link{get_topic_texts}()},
\code{\link{identify_topic_trends}()},
\code{\link{plot_model_comparison}()},
\code{\link{plot_quality_metrics}()},
\code{\link{run_contrastive_topics_internal}()},
\code{\link{run_neural_topics_internal}()},
\code{\link{run_temporal_topics_internal}()}
}
\concept{topic-modeling}
