% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{word_correlation_network}
\alias{word_correlation_network}
\title{Analyze and Visualize Word Correlation Networks}
\usage{
word_correlation_network(
  dfm_object,
  doc_var = NULL,
  common_term_n = 130,
  corr_n = 0.4,
  top_node_n = 40,
  nrows = 1,
  height = 1000,
  width = 900
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{doc_var}{A document-level metadata variable (default: NULL).}

\item{common_term_n}{Minimum number of common terms for filtering terms (default: 30).}

\item{corr_n}{Minimum correlation value for filtering terms (default: 0.4).}

\item{top_node_n}{Number of top nodes to display (default: 40).}

\item{nrows}{Number of rows to display in the table (default: 1).}

\item{height}{The height of the resulting Plotly plot, in pixels (default: 1000).}

\item{width}{The width of the resulting Plotly plot, in pixels (default: 900).}
}
\value{
A list containing the Plotly plot, a data frame of the network layout, and the igraph graph object.
}
\description{
This function creates a word correlation network based on a document-feature matrix (dfm).
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech

  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))

  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")

  dfm_object <- quanteda::dfm(tokens)

  word_correlation_network_results <- TextAnalysisR::word_correlation_network(
                                      dfm_object,
                                      doc_var = "reference_type",
                                      common_term_n = 30,
                                      corr_n = 0.4,
                                      top_node_n = 40,
                                      nrows = 1,
                                      height = 1000,
                                      width = 900)
  print(word_correlation_network_results$plot)
  print(word_correlation_network_results$table)
  print(word_correlation_network_results$summary)
}
}
