% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{word_correlation_network}
\alias{word_correlation_network}
\title{Analyze and Visualize Word Correlation Networks}
\usage{
word_correlation_network(
  dfm_object,
  co_occur_n = 30,
  corr_n = 0.4,
  top_node_n = 40,
  height = 1000,
  width = 900
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{co_occur_n}{Minimum number of co-occurrences for filtering terms (default is 30).}

\item{corr_n}{Minimum correlation value for filtering terms (default is 0.4).}

\item{top_node_n}{Number of top nodes to display (default is 40).}

\item{height}{The height of the resulting Plotly plot, in pixels. Defaults to \code{1000}.}

\item{width}{The width of the resulting Plotly plot, in pixels. Defaults to \code{900}.}
}
\value{
A list containing a Plotly object and a data frame with the results.
}
\description{
This function creates a word correlation network based on a document-feature matrix (dfm).
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
  word_correlation_network_results <- TextAnalysisR::word_correlation_network(
                                      dfm_object,
                                      co_occur_n = 30,
                                      corr_n = 0.4,
                                      top_node_n = 40,
                                      height = 1000,
                                      width = 900)
  word_correlation_network_results$plot
  word_correlation_network_results$table
  word_correlation_network_results$summary
}
}
