% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{word_frequency_trends}
\alias{word_frequency_trends}
\title{Plot Word Frequency Trends Over Time}
\usage{
word_frequency_trends(
  dfm_object,
  stm_model,
  time_variable,
  selected_terms,
  height = 500,
  width = 1000
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{stm_model}{An STM model object.}

\item{time_variable}{The column name for the time variable (e.g., "year").}

\item{selected_terms}{A vector of terms to analyze trends for.}

\item{height}{The height of the resulting Plotly plot, in pixels. Defaults to \code{500}.}

\item{width}{The width of the resulting Plotly plot, in pixels. Defaults to \code{1000}.}
}
\value{
A Plotly object showing interactive word frequency trends over time.
}
\description{
Analyze and visualize word frequency trends over time for a fixed term column.
}
\details{
This function requires a fitted STM model object and a quanteda dfm object.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
  stm_15 <- TextAnalysisR::stm_15
  TextAnalysisR::word_frequency_trends(dfm_object,
                                    stm_model = stm_15,
                                    time_variable = "year",
                                    selected_terms = c("calculator", "computer"),
                                    height = 500,
                                    width = 1000)
}

}
