% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining_functions.R
\name{word_frequency_trends}
\alias{word_frequency_trends}
\title{Analyze and Visualize Trends in Word Frequency}
\usage{
word_frequency_trends(
  dfm_object,
  stm_model,
  continuous_variable,
  selected_terms,
  height = 500,
  width = 900
)
}
\arguments{
\item{dfm_object}{A quanteda document-feature matrix (dfm).}

\item{stm_model}{An STM model object.}

\item{continuous_variable}{A continuous variable in the metadata.}

\item{selected_terms}{A vector of terms to analyze trends for.}

\item{height}{The height of the resulting Plotly plot, in pixels. Defaults to \code{500}.}

\item{width}{The width of the resulting Plotly plot, in pixels. Defaults to \code{900}.}
}
\value{
A list containing a Plotly object and a data frame with the results.
}
\description{
This function analyzes and visualizes trends in word frequency based on a document-feature matrix (dfm) and a fitted STM model.
}
\details{
This function requires a fitted STM model object and a quanteda dfm object.
}
\examples{
if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
  stm_15 <- TextAnalysisR::stm_15
  word_frequency_trends_results <- TextAnalysisR::word_frequency_trends(dfm_object,
                                    stm_model = stm_15,
                                    continuous_variable = "year",
                                    selected_terms = c("calculator", "computer"),
                                    height = 500,
                                    width = 900)
  word_frequency_trends_results$plot
  word_frequency_trends_results$table
}
}
